requirements.txt
yamale~=3.0.8
pyyaml~=6.0
pyyaml-include~=1.2
typing~=3.7.4.3
dataclasses~=0.6
setuptools~=58.0.4